+++ src/library/configs/module.ts
