+++ src/library/core/index.ts
