+++ src/library/core/pipe.dto.ts
